Correccion de funcionalidades

+++ src/main/java/com/busquedaCandidato/candidato/service/CandidateService.java
@@ -1,32 +1,30 @@
 package com.busquedaCandidato.candidato.service;
 
 import com.busquedaCandidato.candidato.dto.request.CandidateRequestDto;
-import com.busquedaCandidato.candidato.dto.response.CandidatePhasesResponseDto;
 import com.busquedaCandidato.candidato.dto.response.CandidateResponse;
 import com.busquedaCandidato.candidato.dto.response.CandidateResponseDto;
-
 import com.busquedaCandidato.candidato.entity.CandidateEntity;
 import com.busquedaCandidato.candidato.entity.PostulationEntity;
 import com.busquedaCandidato.candidato.entity.RoleIDEntity;
 import com.busquedaCandidato.candidato.entity.VacancyCompanyEntity;
-import com.busquedaCandidato.candidato.exception.type.CandidateNoExistException;
-import com.busquedaCandidato.candidato.exception.type.IdCardAlreadyExistException;
+import com.busquedaCandidato.candidato.exception.type.EntityNoExistException;
 import com.busquedaCandidato.candidato.exception.type.PhoneAlreadyExistException;
 import com.busquedaCandidato.candidato.exception.type.RoleIdNoExistException;
-import com.busquedaCandidato.candidato.exception.type.EntityNoExistException;
+import com.busquedaCandidato.candidato.exception.type.CandidateNoExistException;
+import com.busquedaCandidato.candidato.exception.type.IdCardAlreadyExistException;
+import com.busquedaCandidato.candidato.exception.type.EntityAlreadyHasRelationException;
+import com.busquedaCandidato.candidato.mapper.IMapperCandidatePhasesResponse;
 import com.busquedaCandidato.candidato.mapper.IMapperCandidateRequest;
 import com.busquedaCandidato.candidato.mapper.IMapperCandidateResponse;
 import com.busquedaCandidato.candidato.repository.ICandidateRepository;
 import com.busquedaCandidato.candidato.repository.IPostulationRepository;
 import com.busquedaCandidato.candidato.repository.IRoleIDRepository;
 import com.busquedaCandidato.candidato.repository.IVacancyCompanyRepository;
-import com.busquedaCandidato.candidato.specification.CandidateSpecification;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
@@ -41,6 +39,7 @@ public class CandidateService {
     private final IRoleIDRepository roleIDRepository;
     private final IMapperCandidateResponse mapperCandidateResponse;
     private final IMapperCandidateRequest mapperCandidateRequest;
+    private final IMapperCandidatePhasesResponse phasesMappers;
 
     public List<CandidateResponseDto> getCandidateByRole(String roleName) {
 
@@ -111,45 +110,6 @@ public class CandidateService {
         );
     }
 
-    public List<CandidateResponseDto> getSearchCandidatesByNameOrRoleId(String searchValue) {
-        validationQuery(searchValue);
-        List<CandidateEntity> candidates = candidateRepository.findAll(CandidateSpecification.filterBySingleField(searchValue));
-        validationListCandidate(candidates);
-
-        return candidates.stream().map(candidate -> {
-            List<CandidatePhasesResponseDto> phases = candidate.getPostulations().stream()
-                    .flatMap(postulation -> postulation.getProcess().getProcessPhases().stream())
-                    .map(phase -> new CandidatePhasesResponseDto(
-                            phase.getId(),
-                            phase.getProcess().getId(),
-                            phase.getPhase().getId(),
-                            phase.getPhase().getName(),
-                            phase.getState().getId(),
-                            phase.getState().getName(),
-                            phase.getStatus(),
-                            phase.getDescription(),
-                            phase.getAssignedDate(),
-                            phase.getProcess().getPostulation().getVacancyCompany().getId(),
-                            phase.getProcess().getPostulation().getVacancyCompany().getRole().getName(),
-                            phase.getProcess().getPostulation().getVacancyCompany().getJobProfile().getName()
-                    ))
-                    .collect(Collectors.toList());
-
-            return new CandidateResponseDto(
-                    candidate.getId(),
-                    candidate.getName(),
-                    candidate.getLastName(),
-                    candidate.getCard(),
-                    candidate.getBirthdate(),
-                    candidate.getRegistrationDate(),
-                    candidate.getPhone(),
-                    candidate.getCity(),
-                    candidate.getEmail(),
-                    phases
-            );
-        }).collect(Collectors.toList());
-    }
-
     public List<CandidateResponseDto> getByNameCandidate(String name){
         List<CandidateEntity> candidateEntities = candidateRepository.findByName(name);
 
@@ -169,7 +129,9 @@ public class CandidateService {
     }
 
     public List<CandidateResponseDto> getAllCandidate(){
-        return candidateRepository.findAll().stream()
+        List<CandidateEntity> candidates = candidateRepository.findAll();
+
+        return candidates.stream()
                 .map(mapperCandidateResponse::toDto)
                 .collect(Collectors.toList());
     }
@@ -178,6 +140,7 @@ public class CandidateService {
         if(candidateRepository.existsByCard(candidateRequestDto.getCard())){
             throw new IdCardAlreadyExistException();
         }
+
         if(candidateRepository.existsByPhone(candidateRequestDto.getPhone())){
             throw new PhoneAlreadyExistException();
         }
@@ -207,8 +170,13 @@ public class CandidateService {
         CandidateEntity existingCandidate = candidateRepository.findById(id)
                 .orElseThrow(EntityNoExistException::new);
 
+        if (postulationRepository.existsByCandidateId(id)) {
+            throw new EntityAlreadyHasRelationException();
+        }
+
         candidateRepository.delete(existingCandidate);
     }
+
     private void validationQuery(String query){
         if (query == null || query.trim().isEmpty()) {
             throw new IllegalArgumentException("Search query cannot be empty");
