+++ src/main/java/com/busquedaCandidato/candidato/service/JobProfileService.java
@@ -4,10 +4,12 @@ import com.busquedaCandidato.candidato.dto.request.JobProfileRequestDto;
 import com.busquedaCandidato.candidato.dto.response.JobProfileResponseDto;
 import com.busquedaCandidato.candidato.entity.JobProfileEntity;
 import com.busquedaCandidato.candidato.exception.type.EntityAlreadyExistsException;
+import com.busquedaCandidato.candidato.exception.type.EntityAlreadyHasRelationException;
 import com.busquedaCandidato.candidato.exception.type.EntityNoExistException;
 import com.busquedaCandidato.candidato.mapper.IMapperJobProfileRequest;
 import com.busquedaCandidato.candidato.mapper.IMapperJobProfileResponse;
 import com.busquedaCandidato.candidato.repository.IJobProfileRepository;
+import com.busquedaCandidato.candidato.repository.IVacancyCompanyRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import java.util.List;
@@ -17,6 +19,7 @@ import java.util.stream.Collectors;
 @AllArgsConstructor
 public class JobProfileService {
     private final IJobProfileRepository jobProfileRepository;
+    private final IVacancyCompanyRepository vacancyCompanyRepository;
     private final IMapperJobProfileResponse mapperJobProfileResponse;
     private final IMapperJobProfileRequest mapperJobProfileRequest;
 
@@ -56,6 +59,10 @@ public class JobProfileService {
     public void deleteJobProfile(Long id){
         JobProfileEntity existingJob = jobProfileRepository.findById(id)
                 .orElseThrow(EntityNoExistException::new);
+
+        if (vacancyCompanyRepository.existsByJobProfileId(id)) {
+            throw new EntityAlreadyHasRelationException();
+        }
 
         jobProfileRepository.delete(existingJob);
     }
