+++ src/main/java/com/busquedaCandidato/candidato/service/StateService.java
@@ -4,9 +4,11 @@ import com.busquedaCandidato.candidato.dto.request.StateRequestDto;
 import com.busquedaCandidato.candidato.dto.response.StateResponseDto;
 import com.busquedaCandidato.candidato.entity.StateEntity;
 import com.busquedaCandidato.candidato.exception.type.EntityAlreadyExistsException;
+import com.busquedaCandidato.candidato.exception.type.EntityAlreadyHasRelationException;
 import com.busquedaCandidato.candidato.exception.type.EntityNoExistException;
 import com.busquedaCandidato.candidato.mapper.IMapperStateRequest;
 import com.busquedaCandidato.candidato.mapper.IMapperStateResponse;
+import com.busquedaCandidato.candidato.repository.ICandidatePhasesRepository;
 import com.busquedaCandidato.candidato.repository.IStateRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,6 +20,7 @@ import java.util.stream.Collectors;
 public class StateService {
 
     private final IStateRepository stateRepository;
+    private final ICandidatePhasesRepository candidatePhasesRepository;
     private final IMapperStateResponse mapperStateResponse;
     private final IMapperStateRequest mapperStateRequest;
 
@@ -57,6 +60,10 @@ public class StateService {
     public void deleteState(Long id){
         StateEntity existingState = stateRepository.findById(id)
                 .orElseThrow(EntityNoExistException::new);
+
+        if (candidatePhasesRepository.existsByStateId(id)) {
+            throw new EntityAlreadyHasRelationException();
+        }
 
         stateRepository.delete(existingState);
     }
